\name{center}
\alias{center}
\title{Center a variable}
\usage{
  center(x)
}
\arguments{
  \item{x}{Variable to be centered}
}
\value{
  Numeric x - mean(x)
}
\description{
  By centering I mean take the numeric variable and
  subtract the mean from each value.  Then return each
  value.
}
\examples{
center(c(1,2,3))
}

