% Generated by roxygen2 (4.0.1): do not edit by hand
\name{countif}
\alias{countif}
\title{countif}
\usage{
countif(vector, targetval)
}
\arguments{
\item{vector}{The vector in which the items to be counted are being stored}

\item{targetval}{The value that is being compared against}
}
\description{
This is a convenience function that provides the number of entries in the vector that match the target value
}
\examples{
countif(rep(LETTERS,2),"A")
}

