% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleData.R
\name{createSquareMatrix}
\alias{createSquareMatrix}
\title{createSquareMatrix}
\usage{
createSquareMatrix(lowerarray, upperarray, diagarray = NA)
}
\arguments{
\item{lowerarray}{A vector of numeric values (excluding diagonals) that define the lower triangle of the square matrix}

\item{upperarray}{A vector of numeric values (excluding diagonals) that define the lower triangle of the square matrxi}

\item{diagarray}{What you want in the diagonal or a vector that will fill the diagonal}
}
\description{
Given vectors for a square matrix (upper triangle, lower triangle, and diagonal) build the square correlation matrix
}
\examples{
createSquareMatrix(LETTERS[1],LETTERS[6:8],1)
createSquareMatrix(1:3,1,6:8)
createSquareMatrix(1:3,4:6,7:9)
createSquareMatrix(1:6,7:12,13:16)
}

