% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeries.R
\name{fft.binN}
\alias{fft.binN}
\title{Find the number of FFT bins}
\usage{
fft.binN(Nsample)
}
\arguments{
\item{Nsample}{numeric The number of samples in the time series}
}
\description{
Find the number of FFT bins
}

