% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualAngleCalculations.R
\name{in2m}
\alias{in2m}
\title{Convert inches to meters}
\usage{
in2m(inches)
}
\arguments{
\item{inches}{A value in inches}
}
\value{
numeric a value in meters
}
\description{
Convert inches to meters
}

