% Generated by roxygen2 (4.0.1): do not edit by hand
\name{jpeg.save}
\alias{jpeg.save}
\title{Save current plot as a jpeg}
\usage{
jpeg.save(filename)
}
\arguments{
\item{filename}{The filename for the saved plot}

\item{...}{options passed to dev.copy}
}
\description{
Save current plot as a jpeg
}

