% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleStats.R
\name{LScore}
\alias{LScore}
\title{Calculate an L score}
\usage{
LScore(data, lambda)
}
\arguments{
\item{data}{(data.frame) Just the data to L score in the wide format}

\item{lambda}{(numeric) a vector specifying the lambda weights to produce the L scores in the same order as the columns in data}
}
\description{
Calculate an L score
}

