% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualAngleCalculations.R
\name{mi2km}
\alias{mi2km}
\title{Convert miles to kilometers}
\usage{
mi2km(mi)
}
\arguments{
\item{mi}{numeric value in miles (mi)}
}
\value{
numeric a value in km
}
\description{
Convert miles to kilometers
}

