% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleStats.R
\name{minFprime}
\alias{minFprime}
\title{minFprime}
\usage{
minFprime(F1, F2, numdf, F1denomdf, F2denomdf, round = FALSE)
}
\arguments{
\item{F1}{F value from the first test}

\item{F2}{F value from the second test}

\item{numdf}{df of the numerator (the variable of interest)}

\item{F1denomdf}{The df denominator of the first test}

\item{F2denomdf}{The df denominator of the second test}

\item{round}{Whether to round the p value to three places}
}
\description{
Calculates min F' from Clark (1973)
}

