% Generated by roxygen2 (4.0.1): do not edit by hand
\name{padChar}
\alias{padChar}
\title{padChar}
\usage{
padChar(string, padlen, filler = "0", padSide = "left")
}
\arguments{
\item{string}{A single string or vector of strings}

\item{padlen}{The desired length of the padded string}

\item{filler}{A single character that will be used to pad the string}

\item{padSide}{Which side of the string will be padded, default to 'left' all other values are treated as a right side padding.}
}
\description{
Pad a vector of character strings with a single character to a set length
}
\examples{
padChar(c("bob","tool","marks"),5)
padChar(c("bob","tool","marks"),5,"0","right")
}

