% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotContinuousInteraction}
\alias{plotContinuousInteraction}
\title{plotContinuousInteraction}
\usage{
plotContinuousInteraction(lm.obj, DVname, xVar, traceVar,
  traceLvlLabels = c("-1 SD", "Mean", "+1 SD"), xlabel = NULL,
  tracelabel = NULL, ylabel = NULL)
}
\arguments{
\item{lm.obj}{The resulting object from an lm fit (not the summary)}

\item{DVname}{The name of the dependent variable from the model}

\item{xVar}{The variable to be provided on the X axis}

\item{traceVar}{The variable to be represented with different lines}

\item{traceLvlLabels}{The labels for those traces (in order from -1 SD, mean, +1 SD)}

\item{xlabel}{Label for the x-axis}

\item{tracelabel}{Label for the traces}

\item{ylabel}{Label for the y-axis}
}
\description{
Plots a continuous by continuous interaction for a model fit using lm.
}
\details{
This does not create a production ready graph.  It does not handle data with more than two predictor variables.  It is just a shorthand for data-visualization.  The code may provide some suggestions about how to make a nice publication ready graph.
}
\examples{
ex.lm <- lm(Sepal.Length~Sepal.Width*Petal.Length,data=iris[iris$Species=="setosa",])
plotContinuousInteraction(ex.lm,"Sepal.Length","Sepal.Width","Petal.Length")
}

