% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualAngleCalculations.R
\name{radians}
\alias{radians}
\title{Convert degrees to radians}
\usage{
radians(degrees)
}
\arguments{
\item{degrees}{A value in degrees}
}
\value{
The value provided in degrees converted to radians
}
\description{
Convert degrees to radians
}
\examples{
radians(90)
}

