% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{repsych-defunct}
\alias{repsych-defunct}
\alias{t.contrast}
\title{Defunct function(s) in the repsych package}
\description{
These functions were in older versions of the repsych package
or russfnc scripts, but now have been completely removed.
Where possible they will retain a call to .Defunct.
}
\section{Details}{

\tabular{rl}{
  \code{t.contrast} \tab conflicted with the packaging and has been removed, see \code{\link{tContrast}} for a replacement\cr
}
}

