\docType{package}
\name{russmisc-deprecated}
\alias{Conv2Dto3D}
\alias{Conv3Dto2D}
\alias{dist3D.l}
\alias{latinsquare.digram}
\alias{product}
\alias{russmisc-deprecated}
\alias{russmisc-deprecated-package}
\title{Deprecated function(s) in the russmisc package}
\usage{
  latinsquare.digram(...)
}
\arguments{
  \item{...}{Parameters to be passed to the modern version
  of the function}
}
\description{
  These functions are provided for compatiblity with older
  version of the russmisc package and russfnc scripts.
  They may eventually be completely removed.
}
\section{Details}{
  \tabular{rl}{ \code{latinsquare.digram} \tab now a
  synonym for \code{\link{latinSquareDigram}}\cr
  \code{Conv3Dto2D} \tab now a synonym for
  \code{\link{conv3Dto2D}}\cr \code{Conv2Dto3D} \tab now a
  synonym for \code{\link{conv2Dto3D}}\cr \code{dist3D.l}
  \tab now a synonym for \code{\link{dist3D}}\cr
  \code{product} \tab now a synonym for \code{\link{prod}}
  in base\cr }
}

