% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleData.R
\name{SSE}
\alias{SSE}
\title{Calculate the sum of squared errors}
\usage{
SSE(x, y, ...)
}
\arguments{
\item{x}{Numeric data series}

\item{y}{Numeric data series}

\item{...}{further arguments for mean}
}
\value{
numeric
}
\description{
Calculate the sum of squared errors
}

