\docType{data}
\name{StatBingroup}
\alias{StatBingroup}
\title{StatBingroup}
\format{proto object 
 $ calculate_groups:function (., data, ...)   
 $ objname         : chr "bin" 
 $ default_aes     :function (.)   
 $ default_geom    :function (.)   
 $ calculate       :function (., data, scales, binwidth = NULL, origin = NULL, breaks = NULL,  
    width = 0.9, drop = FALSE, right = FALSE, ...)   
 $ required_aes    : chr "x" 
 $ informed        : logi FALSE 
 parent: proto object 
 .. parent: proto object}
\usage{
  StatBingroup
}
\description{
  Analogous to ggplot2:::StatBin
}
\keyword{datasets}

