% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualAngleCalculations.R
\name{viewdist}
\alias{viewdist}
\title{Calculate a viewing distance given a visual angle and stimulus size}
\usage{
viewdist(visangle.val, stimsize.val)
}
\arguments{
\item{visangle.val}{A value in degrees of visual angle}

\item{stimsize.val}{A stimulus size}
}
\value{
The viewing distance of an object of stimsize.val that, when centrally presented, would have a visual angle (in degrees) of visangle.val.  Viewing distance is reported in the same units as stimsize.val.
}
\description{
Calculate a viewing distance given a visual angle and stimulus size
}
\examples{
viewdist(visangle(10,100),10)
}
\author{
Russell S. Pierce \email{Russell.S.Pierce@gmail.com}
}
\seealso{
\code{\link{viewdist} \link{visangle}}
}

