% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualAngleCalculations.R
\name{visangle3D.l}
\alias{visangle3D.l}
\title{Helper function for visangle3D to handle lists}
\usage{
visangle3D.l(viewerloc, stimloc, stimdim)
}
\arguments{
\item{viewerloc}{The location of the center of the viewer as a list}

\item{stimloc}{The location of the center of the stimulus as a list}

\item{stimdim}{The stimulus dimensions as a numeric list}
}
\description{
Helper function for visangle3D to handle lists
}

