% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleStats.R
\name{Zmodified}
\alias{Zmodified}
\title{Calculate the Iglewicz and Hoaglin (1993) Z modified}
\usage{
Zmodified(x)
}
\arguments{
\item{x}{A set of numeric scores}
}
\description{
Calculate the Iglewicz and Hoaglin (1993) Z modified
}

